requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
newsapi-python==0.2.6
alpha_vantage==2.3.1
matplotlib==3.5.1
seaborn==0.11.2 